pyproject.toml
[project]
name = "agent-workflow"
version = "0.1.0"
description = "Manus-style whiteboard analysis workflow service"
readme = "specs/001-manus-analysis-workflow/quickstart.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "jinja2>=3.1.0,<4.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pydantic-settings>=2.3.0,<3.0.0",
  "pyyaml>=6.0.0,<7.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "mypy>=1.11.0,<2.0.0",
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "ruff>=0.6.0,<1.0.0",
  "schemathesis>=3.34.0,<4.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
